requirements.txt
asgiref==3.3.1
certifi~=2019.11.28
chardet~=3.0.4
click==7.1.2
contextvars==2.4
Django~=3.1.4
django-cors-headers==3.6.0
django-redis==4.12.1
djangorestframework~=3.11.0
elasticsearch==7.11.0
gunicorn==20.0.4
h11==0.11.0
httpcore==0.12.2
httptools==0.1.1
httpx==0.16.1
idna~=2.8
immutables==0.15
pkg-resources==0.0.0
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
rfc3986==1.4.0
sniffio==1.2.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3~=1.25.11
uvicorn==0.13.2
whitenoise==5.2.0
sentry-sdk==0.20.3

python-dateutil~=2.8.1
requests~=2.22.0
pip~=20.0.2
keyring~=18.0.1
colorama~=0.4.3
PyYAML~=5.3.1
numpy~=1.19.4
docutils~=0.15.2
uvloop~=0.14.0
cryptography~=2.8
six~=1.14.0
setuptools~=45.2.0
simplejson~=3.16.0
Markdown~=3.1.1
Pygments~=2.3.1
gobiko.apns~=0.1.6
